+++ libraries/session/src/main/java/androidx/media3/session/LegacyConversions.java
@@ -1369,17 +1369,16 @@ import java.util.concurrent.TimeoutException;
         || hasAction(actions, PlaybackStateCompat.ACTION_PLAY_PAUSE)) {
       playerCommandsBuilder.add(COMMAND_PLAY_PAUSE);
     }
-    if (hasAction(actions, PlaybackStateCompat.ACTION_PREPARE)) {
-      playerCommandsBuilder.add(COMMAND_PREPARE);
+    if (hasAction(actions, PlaybackStateCompat.ACTION_PLAY_FROM_MEDIA_ID)
+        || hasAction(actions, PlaybackStateCompat.ACTION_PLAY_FROM_SEARCH)
+        || hasAction(actions, PlaybackStateCompat.ACTION_PLAY_FROM_URI)) {
+      playerCommandsBuilder.add(COMMAND_SET_MEDIA_ITEM);
     }
-    if ((hasAction(actions, PlaybackStateCompat.ACTION_PREPARE_FROM_MEDIA_ID)
-            && hasAction(actions, PlaybackStateCompat.ACTION_PLAY_FROM_MEDIA_ID))
-        || (hasAction(actions, PlaybackStateCompat.ACTION_PREPARE_FROM_SEARCH)
-            && hasAction(actions, PlaybackStateCompat.ACTION_PLAY_FROM_SEARCH))
-        || (hasAction(actions, PlaybackStateCompat.ACTION_PREPARE_FROM_URI)
-            && hasAction(actions, PlaybackStateCompat.ACTION_PLAY_FROM_URI))) {
-      // Require both PREPARE and PLAY actions as we have no logic to handle having just one action.
-      playerCommandsBuilder.addAll(COMMAND_SET_MEDIA_ITEM, COMMAND_PREPARE);
+    if (hasAction(actions, PlaybackStateCompat.ACTION_PREPARE)
+        || hasAction(actions, PlaybackStateCompat.ACTION_PREPARE_FROM_MEDIA_ID)
+        || hasAction(actions, PlaybackStateCompat.ACTION_PREPARE_FROM_SEARCH)
+        || hasAction(actions, PlaybackStateCompat.ACTION_PREPARE_FROM_URI)) {
+      playerCommandsBuilder.add(COMMAND_PREPARE);
     }
     if (hasAction(actions, PlaybackStateCompat.ACTION_REWIND)) {
       playerCommandsBuilder.add(COMMAND_SEEK_BACK);